Revise code to age standardize mortality rates and generate graphs
The causes of death in the original mortality processing were revised to have one version with the SIAMH causes of death (with istroke) and one version with the more detailed causes that were used for the decomposition.

The code for the protocol paper was then revised to have one code generating weights and one code generating age standardized rates and respective graphs.