Update income function for 2021 data
I updated the recode_income function in 1_processing_functions.R since the income variable in the 2021 BRFSS data ("INCOME3) was expanded compared to previous years.

I added a comment in the 1.process_raw_data.R script since line 24 has to be changed as soon as the 2021 data is permantly added.